data types have been changed from int to uint16_t

+++ components/accelerometer/accelerometer.h
@@ -28,8 +28,9 @@
 
 void accel_config(spi_device_handle_t *spi);
 void read_acceleration_task(void* pvParameters);
-void pwm_leds(int duty, int i);
-int chromatic_mode(int accel_data,  char **note);
-int pentatonic_mode(int accel_data, char **note);
+void pwm_leds(int16_t  duty, int16_t i);
+int chromatic_mode(int16_t accel_data,  char **note);
+int pentatonic_mode(int16_t accel_data, char **note);
 void notes(int16_t freq, int16_t  duty);
+
 #endif
